Phpstan level 3

+++ src/Util/Temporal.php
@@ -405,7 +405,7 @@ class Temporal
 	 */
 	public static function getDaysInMonth(int $y, int $m): int
 	{
-		return date('t', mktime(0, 0, 0, $m, 1, $y));
+		return (int) date('t', mktime(0, 0, 0, $m, 1, $y));
 	}
 
 	/**
